requirements.txt
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1
colorama==0.4.6
tabulate==0.9.0
click==8.1.7
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0